+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Credential service provider integrated
+- Custom retry attempts and time for transactions waiting functionality
 
 ## [0.0.3-alpha] - 2023-01-31